Add methods to Node
